divided JSON based tests from database query tests
